fix: repair remote proxy CI failures
